fix optimizer tasks constructions (#4261)
